+++ src/main/java/racingcar/service/RacingService.java
@@ -1,9 +1,5 @@
 package racingcar.service;
 
-import static racingcar.exception.ExceptionMessage.INVALID_INPUT_NUMBER;
-import static racingcar.exception.ExceptionMessage.MINIMUM_RACING_TIME;
-import static racingcar.exception.ExceptionMessage.USER_INPUT_NOT_NULL;
-
 import java.util.List;
 import racingcar.domain.Car;
 import racingcar.domain.CarCreator;
@@ -11,19 +7,20 @@ import racingcar.domain.CarManager;
 import racingcar.domain.InputParser;
 import racingcar.domain.Racing;
 import racingcar.utils.RandomNumber;
+import racingcar.validation.Validator;
 
 
 public class RacingService {
 
-    private static final String CHECK_NUMBER_REGEX = "\\d+";
-
-    private CarManager carManager;
     private final Racing racing;
     private final RandomNumber randomNumber;
+    private final Validator validation;
+    private CarManager carManager;
 
-    public RacingService( Racing racing, RandomNumber randomNumber){
+    public RacingService(Racing racing, RandomNumber randomNumber, Validator validation) {
         this.racing = racing;
         this.randomNumber = randomNumber;
+        this.validation = validation;
     }
 
     public void setUpRacing(String userInput) {
@@ -43,30 +40,14 @@ public class RacingService {
         return racing.findWinner(carManager.getCars());
     }
 
-
     public String getValidatedStringInput(String userStringInput) {
-        verifyUserInput(userStringInput);
+        validation.verifyUserInput(userStringInput);
         return userStringInput;
     }
 
     public int getValidateIntegerInput(String userIntegerInput) {
-        verifyRepeatCount(userIntegerInput);
+        validation.verifyRepeatCount(userIntegerInput);
         return Integer.parseInt(userIntegerInput);
-    }
-
-    private void verifyRepeatCount(String userInput) {
-        if (!userInput.matches(CHECK_NUMBER_REGEX)) {
-            throw new IllegalArgumentException(INVALID_INPUT_NUMBER.getMessage());
-        }
-        if (Integer.parseInt(userInput) == 0) {
-            throw new IllegalArgumentException(MINIMUM_RACING_TIME.getMessage());
-        }
-    }
-
-    private void verifyUserInput(String userInput) {
-        if (userInput.isBlank() || userInput == null) {
-            throw new IllegalArgumentException(USER_INPUT_NOT_NULL.getMessage());
-        }
     }
 
 }
